Check permissions when deleting comments

+++ portfolio/src/main/java/com/google/sps/servlets/CommentDeleteAllServlet.java
@@ -30,31 +30,40 @@ import com.google.appengine.api.datastore.PreparedQuery;
 import com.google.appengine.api.datastore.Query;
 import com.google.appengine.api.datastore.Transaction;
 import com.google.appengine.api.datastore.TransactionOptions;
+import com.google.appengine.api.users.UserService;
+import com.google.appengine.api.users.UserServiceFactory;
 
 /** Servlet that returns comments. */
 @WebServlet("/commentDeleteAll")
 public class CommentDeleteAllServlet extends HttpServlet {
   @Override
   public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
+    UserService userService = UserServiceFactory.getUserService();
+
+    if (!userService.isUserAdmin()) {
+      response.sendError(HttpServletResponse.SC_FORBIDDEN, "admin required");
+      return;
+    }
+
     DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
 
     // Transaction doesn't support try-with-resources??
     Transaction txn = datastore.beginTransaction(
-        TransactionOptions.Builder.withXG(true)
+      TransactionOptions.Builder.withXG(true)
     );
 
     try {
-        Query query = new Query("comment");
-        PreparedQuery results = datastore.prepare(query);
+      Query query = new Query("comment");
+      PreparedQuery results = datastore.prepare(query);
 
-        for (Entity entity : results.asIterable()) {
-            datastore.delete(txn, entity.getKey());
-        }
-        txn.commit();
+      for (Entity entity : results.asIterable()) {
+          datastore.delete(txn, entity.getKey());
+      }
+      txn.commit();
     } finally {
-        if (txn.isActive()) {
-            txn.rollback();
-        }
+      if (txn.isActive()) {
+          txn.rollback();
+      }
     }
 
     response.setStatus(HttpServletResponse.SC_NO_CONTENT);
